requirements.txt
streamlit
streamlit-webrtc
av
google-genai
numpy
scipy
matplotlib
pydub
aiortc
opencv-python-headless
python-dotenv
